+++ src/main/java/miu/edu/cs545waa/service/ProductCategoryServiceImpl.java
@@ -14,7 +14,7 @@ public class ProductCategoryServiceImpl implements ProductCategoryService{
     ProductCategoryRepo productCategoryRepo;
 
     @Override
-    public ProductCategory getCategoryById(int id) {
+    public ProductCategory getCategoryById(Long id) {
         return productCategoryRepo.findById(id).get();
     }
